// pweb-study-and-sync/src/app/core/interfaces/user.register.interface.ts
export interface UserRegisterData {
    nameRegisterField: string;
    emailRegisterField: string;
    passwordRegisterField: string;
    repeat_passwordRegisterField: string;
    statusRegisterField: string;
    birthdayRegisterField: string;
    disciplineRegisterField: string;
    qualificationRegisterField: string ;
    semesterRegisterField: string;
  }